Reorganized components.
Put each component into its own file.
