+++ src/Ball.js
@@ -1,0 +1,7 @@
+function Ball(props) {
+  return (
+      <div className='ball' style={{ top: props.top + 'px', left: props.left + 'px' }} />
+  );
+}
+
+export default Ball;
