+++ src/Result.js
@@ -1,0 +1,9 @@
+function Result(props) {
+  return (
+      <div className='result'>
+          <h1>{props.winMessage}</h1>
+      </div>
+  );
+}
+
+export default Result;
